magento/adobe-stock-integration#14: Implement content type search filter --refactoring
