--- AdobeStockImageAdminUi/Ui/Component/Listing/Columns/ContentType/Illustrations/Options.php
@@ -1,29 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\AdobeStockImageAdminUi\Ui\Component\Listing\Columns\ContentType\Illustrations;
-
-/**
- * Content Type Illustrations filter options provider
- */
-class Options implements \Magento\Framework\Data\OptionSourceInterface
-{
-    /**
-     * Get options
-     *
-     * @return array
-     */
-    public function toOptionArray(): array
-    {
-        return [
-            [
-                'value' => 1,
-                'label' => __('Yes')
-            ]
-        ];
-    }
-}
